fix logic issues in AddScreen, change type from Int to long

+++ app/src/main/java/es/infolojo/newkeepitdroid/ui/screens/update/UpdateScreen.kt
@@ -43,9 +43,8 @@ fun UpdateScreen(
     isPreview: Boolean = false,
     viewModel: UpdateScreenViewModel? = hiltViewModel(),
     mainEvents: (MainEvents) -> Unit = {},
-    noteId: Int
+    noteId: Long
 ) {
-    // needed values
     // launch the viewModel
     viewModel?.init(noteId, mainEvents)
     // estructura / esqueleto
